Second Push
Recreated the MyStore project and have been working on several components.

Currently the project will list the products on the page and the issue outstanding is the routes not working.

The cart service appears to be working but I'll confirm that more once the routes issues is resolved and I can then begin work on the cart component